+++ src/test/java/lotto/model/lotto/result/RateOfReturnTest.java
@@ -1,4 +1,4 @@
-package lotto.model.result;
+package lotto.model.lotto.result;
 
 import org.junit.jupiter.api.DisplayName;
